apps/backend/pyproject.toml
[project]
name = "decisio-backend"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.30.6",

  "pydantic==2.9.2",
  "pydantic-settings==2.5.2",
  "email-validator==2.2.0",

  "SQLAlchemy==2.0.36",
  "psycopg[binary]==3.2.3",
  "alembic==1.13.3",

  "python-jose==3.3.0",
  "passlib==1.7.4",

  "python-multipart==0.0.12",

  "prometheus-client==0.21.0",
]